user controls file folder restructuring
