+++ CID_Tester/View/Controls/AddTestPlan/TestPlanForm.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="CID_Tester.View.Controls.AddTestPlanControl"
+﻿<UserControl x:Class="CID_Tester.View.Controls.AddTestPlan.TestPlanForm"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View.Controls"
+             xmlns:local="clr-namespace:CID_Tester.View.Controls.AddTestPlan"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <StackPanel Margin="20">
