+++ CID_Tester/View/Controls/Dashboard/DashboardMetricControl.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="CID_Tester.View.Controls.DashboardMetricControl"
+﻿<UserControl x:Class="CID_Tester.View.Controls.Dashboard.DashboardMetricControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View.Controls"
+             xmlns:local="clr-namespace:CID_Tester.View.Controls.Dashboard"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <StackPanel>
