+++ CID_Tester/View/Controls/Dashboard/TestPlanStatus.xaml
@@ -1,17 +1,16 @@
-﻿<UserControl x:Class="CID_Tester.View.Controls.TestPlanStatusControl"
+﻿<UserControl x:Class="CID_Tester.View.Controls.Dashboard.TestPlanStatus"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View.Controls"
+             xmlns:local="clr-namespace:CID_Tester.View.Controls.Dashboard"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-
     <StackPanel HorizontalAlignment="Center">
         <TextBlock>
-        <Hyperlink Command="{Binding ImportCsvCommand}">Import .csv file</Hyperlink>
-        or 
-        <Hyperlink Command="{Binding NavigateToTestPlanCommand}">Create new Test Plan</Hyperlink>
+    <Hyperlink Command="{Binding ImportCsvCommand}">Import .csv file</Hyperlink>
+    or 
+    <Hyperlink Command="{Binding NavigateToTestPlanCommand}">Create new Test Plan</Hyperlink>
         </TextBlock>
     </StackPanel>
 </UserControl>
