--- CID_Tester/View/Controls/TestParameterPropertiesControl.xaml
@@ -1,35 +1,0 @@
-﻿<UserControl x:Class="CID_Tester.View.Controls.TestParameterPropertiesControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View.Controls"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="250">
-    <ScrollViewer VerticalScrollBarVisibility="Auto">
-        <StackPanel Grid.Row="0" Margin="10">
-            <TextBlock Margin="0 0 0 20">Main</TextBlock>
-            <StackPanel Margin="20 0 0 10" Orientation="Horizontal">
-                <TextBlock Margin="0 0 10 0" MinWidth="40">Name</TextBlock>
-                <TextBox Text="{Binding Name}" Width="140" KeyDown="TextBox_KeyDown" />
-            </StackPanel>
-            <StackPanel Margin="20 0 0 10" Orientation="Horizontal">
-                <TextBlock Margin="0 0 10 0" MinWidth="40">Metric</TextBlock>
-                <TextBox Text="{Binding Metric}" Width="140" KeyDown="TextBox_KeyDown" />
-            </StackPanel>
-            <StackPanel Margin="20 0 0 20" Orientation="Horizontal">
-                <TextBlock Margin="0 0 10 0" Width="40">Target</TextBlock>
-                <TextBox Text="{Binding Target}" Width="140" KeyDown="TextBox_KeyDown" HorizontalAlignment="Stretch" />
-            </StackPanel>
-            <StackPanel Margin="20 0 20 10" Orientation="Vertical">
-                <TextBlock Margin="0 0 10 10" MinWidth="40">Description</TextBlock>
-                <TextBox 
-                    Text="{Binding Description}" 
-                    HorizontalAlignment="Stretch"
-                    AcceptsReturn="True" 
-                    KeyDown="TextBox_KeyDown" 
-                    Height="100"/>
-            </StackPanel>
-        </StackPanel>
-    </ScrollViewer>
-</UserControl>
