+++ CID_Tester/View/Controls/TestPlan/TestParameterTable.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="CID_Tester.View.Controls.TestParameterTableControl"
+﻿<UserControl x:Class="CID_Tester.View.Controls.TestPlan.TestParameterTable"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View.Controls"
+             xmlns:local="clr-namespace:CID_Tester.View.Controls.TestPlan"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <StackPanel>
@@ -14,9 +14,9 @@
         <DataGrid ItemsSource="{Binding TestParameters}" SelectedValue="{Binding SelectedTestParameterId, Mode=TwoWay}" SelectedValuePath="ParamCode">
             <DataGrid.InputBindings>
                 <MouseBinding 
-                    Gesture="LeftDoubleClick"
-                    Command="{Binding DoubleClickCommand}"
-                    CommandParameter="{Binding TestParameters/}"/>
+                Gesture="LeftDoubleClick"
+                Command="{Binding DoubleClickCommand}"
+                CommandParameter="{Binding TestParameters/}"/>
             </DataGrid.InputBindings>
             <DataGrid.Columns>
                 <DataGridTextColumn Header="Name" Binding="{Binding Name}" MinWidth="200" />
