--- CID_Tester/View/Controls/TestPlanStatusControl.xaml.cs
@@ -1,14 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace CID_Tester.View.Controls;
-
-/// <summary>
-/// Interaction logic for TestPlanStatusControl.xaml
-/// </summary>
-public partial class TestPlanStatusControl : UserControl
-{
-    public TestPlanStatusControl()
-    {
-        InitializeComponent();
-    }
-}
