+++ CID_Tester/View/Windows/OpenTestPlanView.xaml
@@ -4,8 +4,8 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:CID_Tester.View.Windows"
-        xmlns:controls="clr-namespace:CID_Tester.View.Controls"
-        xmlns:vm="clr-namespace:CID_Tester.ViewModel.Controls"
+        xmlns:AddTestPlan="clr-namespace:CID_Tester.View.Controls.AddTestPlan"
+        xmlns:vm="clr-namespace:CID_Tester.ViewModel.Controls.AddTestPlan"
         mc:Ignorable="d"
         Title="Open Test Plan" Height="600" Width="950" Background="#333334">
     <Window.Resources>
@@ -31,13 +31,13 @@
         </Style>
 
         <DataTemplate DataType="{x:Type vm:AddTestPlanViewModel}">
-            <controls:AddTestPlanControl />
+            <AddTestPlan:TestPlanForm />
         </DataTemplate>
         <DataTemplate DataType="{x:Type vm:AddTestPlanTableSelectorViewModel}">
-            <controls:AddTestPlanTableSelector />
+            <AddTestPlan:TestPlanTable />
         </DataTemplate>
         <DataTemplate DataType="{x:Type vm:AddTestPlanImporterViewModel}">
-            <controls:AddTestPlanImporter />
+            <AddTestPlan:TestPlanImporter />
         </DataTemplate>
     </Window.Resources>
     <Grid>
