+++ CID_Tester/ViewModel/Controls/DashboardMetricViewModel.cs
