+++ CID_Tester/ViewModel/Document/TestPlanStatusControl.cs
@@ -1,0 +1,8 @@
+﻿using System.Windows.Controls;
+
+namespace CID_Tester.ViewModel.Document
+{
+    internal class TestPlanStatusControl : UserControl
+    {
+    }
+}
