+++ CID_Tester/ViewModel/Windows/OpenTestPlanViewModel.cs
@@ -1,7 +1,7 @@
 ﻿
 using CID_Tester.Model;
 using CID_Tester.ViewModel.Command;
-using CID_Tester.ViewModel.Controls;
+using CID_Tester.ViewModel.Controls.AddTestPlan;
 using System.Windows.Input;
 
 namespace CID_Tester.ViewModel.Windows;
